+++ typescript/infra/src/config/chain.ts
@@ -1,6 +1,8 @@
+import { SecretManagerServiceClient } from '@google-cloud/secret-manager';
 import { providers } from 'ethers';
 
 import {
+  ChainMap,
   ChainMetadata,
   ChainName,
   HyperlaneSmartProvider,
@@ -32,6 +34,7 @@ export async function fetchProvider(
   const single = connectionType === RpcConsensusType.Single;
   let rpcData = chainMetadata.rpcUrls.map((url) => url.http);
   if (rpcData.length === 0) {
+    // todo should probably come back here
     rpcData = await getSecretRpcEndpoint(environment, chainName, !single);
   }
 
@@ -72,4 +75,46 @@ export function getChainMetadatas(chains: Array<ChainName>) {
   );
 
   return { ethereumMetadatas, nonEthereumMetadatas };
+}
+
+export async function getSecretMetadataOverrides(
+  deployEnv: DeployEnvironment,
+  chains: string[],
+): Promise<ChainMap<Partial<ChainMetadata>>> {
+  const projectId = 'abacus-labs-dev';
+
+  const client = new SecretManagerServiceClient({
+    projectId,
+  });
+
+  const chainMetadataOverrides: ChainMap<Partial<ChainMetadata>> = {};
+
+  for (const chain of chains) {
+    const secretName = `${deployEnv}-rpc-endpoints-${chain}`;
+    const [secretVersion] = await client.accessSecretVersion({
+      name: `projects/${projectId}/secrets/${secretName}/versions/latest`,
+    });
+    const secretData = secretVersion.payload?.data;
+    if (!secretData) {
+      console.warn('Secret missing payload', secretName);
+      continue;
+    }
+
+    // Handle both string and Uint8Array
+    let dataStr: string;
+    if (typeof secretData === 'string') {
+      dataStr = secretData;
+    } else {
+      dataStr = new TextDecoder().decode(secretData);
+    }
+
+    const rpcUrls = JSON.parse(dataStr);
+    chainMetadataOverrides[chain] = {
+      rpcUrls: rpcUrls.map((rpcUrl: string) => ({
+        http: rpcUrl,
+      })),
+    };
+  }
+
+  return chainMetadataOverrides;
 }
