+++ typescript/infra/test/agent-configs.test.ts
@@ -10,7 +10,6 @@ import { supportedChainNames as testnet4SupportedChainNames } from '../config/en
 import { getAgentConfigJsonPath } from '../scripts/agent-utils.js';
 import { ensureAgentChainConfigIncludesAllChainNames } from '../src/config/agent/agent.js';
 import { AgentEnvironment } from '../src/config/environment.js';
-import { GcpSecretRpcUrlRegistry } from '../src/gcp-registry.js';
 import { readJSONAtPath } from '../src/utils/utils.js';
 
 const environmentChainConfigs = {
@@ -54,10 +53,5 @@ describe('Agent configs', () => {
         );
       });
     });
-  });
-
-  it('testing gcp registry', async () => {
-    const gcpRegistry = new GcpSecretRpcUrlRegistry('mainnet3', rootLogger);
-    await gcpRegistry.listRegistryContent();
   });
 });
